Move multitap out of tap.dart
Tap and the othes classes in tap.dart are not really related.
